Make BatchMakerPass a ModulePass
When we change the Function prototype, we are changing the
current Function type. In LLVM, updating already created
Function's type is not allowed, hence new function need to
be created and added to the module. This can be done only
in module pass.
This commit contains all the necessary modifications to
make it module pass.
We identify functions we need to work on in a module pass
and do our transformation there.