Check use of MemorySSA
Check the possibility of using memorySSA for ease of analysis.
Although it gives direct access to all the memory usage, still
How useful this is for applying in our case is not clear.
I don't see very clear advantage of this - getting all
memory defs and uses.

Idea was to use this for analysing the cache usage.